requirements.txt
# Core dependencies for RAG system
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0
